incVersion.sh: Use committed date instead authored date for timestamp
If a commit is pushed with an authored date in the past the resulting
package might be ignored as old even though it may include the latest
code available.

Example:

The following commit was pushed to ngfw_pkgs:

https://github.com/untangle/ngfw_pkgs/commit/93b355318a7021581a46222c20c758eb0e26e455

Its authored date is: Fri Mar 23 15:13:55 2018

But it was pushed to github on (its committed date): Wed Mar 28 09:19:05 2018

Pushing the commit generated a build and the untangle-sync-settings package
was built successfully:

http://build-master-stretch.untangle.int:8010/builders/amd64-build-master_stretch/builds/1125

The package was labeled:

14.0.0.20180323T1213550700.93b355318-1stretch

93b355318 was the commit hash for the latest commit, but the date that
precedes it was 5 days in the past.

Doing an 'apt-get update; apt-cache show untangle-sync-settings' after this
build still yielded 14.0.0.20180328T0420040700.64aabf0ab-1stretch as the
latest untangle-sync-settings package available because its date is after
that of the latest package built.

If the intent is for the latest package available to refer to that package
built from the latest code committed to the github repository, then the
timestamp should be generated from the commit date of the commit instead of
the author date.