In non-verbose mode, background the build process and actively wait for it
The Travis conundrum:

  - if we build the kernel in VERBOSE mode, the output grows larger than
  4M and Travis forcibly terminates our job (see
  https://docs.travis-ci.com/user/common-build-problems/#log-length-exceeded).
  Raising that limit on the Travis side will simply not happen: see for
  instance https://github.com/travis-ci/travis-ci/issues/1382, which is
  "wont-fix".

  - if we build it without VERBOSE mode, then there is not output for a
  very long time... and Travis forcibly terminates our job (see
  https://docs.travis-ci.com/user/common-build-problems/#build-times-out-because-no-output-was-received).
  Travis does provide a travis-wait() wrapper to avoid this problem, but
  we really insist on having pkgtools remain platform-agnostic: enabling
  developers to run builds on their own machines, but in conditions and
  environments *exactly* similar to our official builds, is a key
  functionality.

We therefore choose to implement our own simple wait-for-pid() wrapper,
based on ps(1), and we change the non-VERBOSE build implementation from
"foreground run" to "background run coupled with wait-for-pid".

Changing the core semantics of build.sh at this point is really
unfortunate, because it's been stable for a good while now and was
giving good result for the rest of our Travis-based builds for userland
packages, but it cannot be helped.

Comment: #12817